backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-filter==24.2
drf-spectacular==0.27.2
gunicorn==22.0.0
google-auth==2.29.0
requests==2.32.3
